+++ index.js
@@ -5,6 +5,7 @@ ENV.co = require('./src/co')
 ENV.li = require('./src/li')
 ENV.tx = require('./src/text')
 ENV.cm = require('./src/comment')
+ENV.fr = require('./src/fragment')
 ENV.namespaces = require('./src/namespaces')
 
 module.exports = ENV